backend/requirements.txt
Flask
python-dotenv
requests
beautifulsoup4
google-generativeai
pdfminer.six